added authorization to controllers

+++ FastFood.Application/Restaurant/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,16 +1,23 @@
 ﻿using Domain.Domain.Exceptions;
+using FastFood.Application.Authorization;
+using FastFood.Domain.Exceptions;
 using FastFood.Domain.Interfaces;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 
 namespace FastFood.Application.Restaurant.Commands.DeleteRestaurant
 {
     public class DeleteRestaurantCommandHandler : IRequestHandler<DeleteRestaurantCommand>
     {
         private readonly IRestaurantRepository _repository;
+        private readonly IAuthorizationService _authorization;
+        private readonly IUserContextService _userContext;
 
-        public DeleteRestaurantCommandHandler(IRestaurantRepository repository)
+        public DeleteRestaurantCommandHandler(IRestaurantRepository repository, IAuthorizationService authorization, IUserContextService userContext)
         {
             _repository = repository;
+            _authorization = authorization;
+            _userContext = userContext;
         }
 
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -20,6 +27,11 @@ namespace FastFood.Application.Restaurant.Commands.DeleteRestaurant
             if (restaurant == null)
             {
                 throw new NotFoundException("Restaurant not found");
+            }
+            var authorizationResult = await _authorization.AuthorizeAsync(_userContext.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete));
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbiddenException();
             }
 
             await _repository.Delete(restaurant);
